refactor - extract runtime ext builder, Zinna api builder
